reword advanced options, break up project payer modal component
